# feat: implement query optimization with 99% performance gains