… src/scripts/test-query-optimization.ts
… src/services/query-cache.service.ts
… src/utils/query-builder.utils.ts